requirements.txt
pandas
openpyxl
psycopg2-binary
pymodbus
tqdm
